refactor(MemBlock): move MemBlock.scala from backend to mem
