+++ src/main/scala/xiangshan/backend/dispatch/Dispatch.scala
@@ -22,13 +22,13 @@ import chisel3.util._
 import difftest._
 import utils._
 import utility._
-import xiangshan.ExceptionNO._
 import xiangshan._
-import xiangshan.backend.MemCoreTopDownIO
+import xiangshan.ExceptionNO._
 import xiangshan.backend.rob.{RobDispatchTopDownIO, RobEnqIO}
-import xiangshan.mem.mdp._
 import xiangshan.backend.Bundles.DynInst
 import xiangshan.backend.fu.FuType
+import xiangshan.mem.mdp._
+import xiangshan.mem.MemCoreTopDownIO
 
 case class DispatchParameters
 (
