+++ src/main/scala/xiangshan/mem/MemBlock.scala
@@ -14,7 +14,7 @@
 * See the Mulan PSL v2 for more details.
 ***************************************************************************************/
 
-package xiangshan.backend
+package xiangshan.mem
 
 import org.chipsalliance.cde.config.Parameters
 import chisel3._
@@ -24,29 +24,31 @@ import freechips.rocketchip.diplomacy.{BundleBridgeSource, LazyModule, LazyModul
 import freechips.rocketchip.interrupts.{IntSinkNode, IntSinkPortSimple}
 import freechips.rocketchip.tile.HasFPUParameters
 import freechips.rocketchip.tilelink._
-import coupledL2.{PrefetchRecv}
 import device.MsiInfoBundle
 import utils._
 import utility._
+import system.SoCParamsKey
 import xiangshan._
+import xiangshan.ExceptionNO._
+import xiangshan.frontend.HasInstrMMIOConst
 import xiangshan.backend.Bundles.{DynInst, MemExuInput, MemExuOutput}
 import xiangshan.backend.ctrlblock.{DebugLSIO, LsTopdownInfo}
 import xiangshan.backend.exu.MemExeUnit
 import xiangshan.backend.fu._
 import xiangshan.backend.fu.FuType._
-import xiangshan.backend.rob.{RobDebugRollingIO, RobPtr}
 import xiangshan.backend.fu.util.{HasCSRConst, SdtrigExt}
-import xiangshan.cache._
-import xiangshan.cache.mmu._
+import xiangshan.backend.{BackendToTopBundle, TopToBackendBundle}
+import xiangshan.backend.rob.{RobDebugRollingIO, RobPtr, RobLsqIO}
+import xiangshan.backend.datapath.NewPipelineConnect
+import xiangshan.backend.fu.NewCSR.{CsrTriggerBundle, TriggerUtil}
+import xiangshan.backend.trace.{Itype, TraceCoreInterface}
+import xiangshan.backend.Bundles._
 import xiangshan.mem._
 import xiangshan.mem.mdp._
-import xiangshan.frontend.HasInstrMMIOConst
 import xiangshan.mem.prefetch.{BasePrefecher, L1Prefetcher, SMSParams, SMSPrefetcher}
-import xiangshan.backend.datapath.NewPipelineConnect
-import system.SoCParamsKey
-import xiangshan.backend.fu.NewCSR.TriggerUtil
-import xiangshan.ExceptionNO._
-import xiangshan.backend.trace.{Itype, TraceCoreInterface}
+import xiangshan.cache._
+import xiangshan.cache.mmu._
+import coupledL2.{PrefetchRecv}
 
 trait HasMemBlockParameters extends HasXSParameter {
   // number of memory units
