+++ src/main/scala/xiangshan/mem/lsqueue/LoadQueueUncache.scala
@@ -28,7 +28,7 @@ import utility._
 import xiangshan.backend.Bundles
 import xiangshan.backend.Bundles.{DynInst, MemExuOutput}
 import xiangshan.backend.fu.FuConfig.LduCfg
-import xiangshan.backend.HasMemBlockParameters
+import xiangshan.mem.HasMemBlockParameters
 
 class UncacheEntry(entryIndex: Int)(implicit p: Parameters) extends XSModule
   with HasCircularQueuePtrHelper
@@ -69,13 +69,13 @@ class UncacheEntry(entryIndex: Int)(implicit p: Parameters) extends XSModule
   val uncacheState = RegInit(s_idle)
   val uncacheData = Reg(io.uncache.resp.bits.data.cloneType)
   val nderr = RegInit(false.B)
-  
+
   val writeback = Mux(req.nc, io.ncOut.fire, io.mmioOut.fire)
   val slaveAck = req_valid && io.uncache.idResp.valid && io.uncache.idResp.bits.mid === entryIndex.U
 
   /**
     * Flush
-    * 
+    *
     * 1. direct flush during idle
     * 2. otherwise delayed flush until receiving uncache resp
     */
@@ -289,7 +289,7 @@ class LoadQueueUncache(implicit p: Parameters) extends XSModule
     // exception generated by outer bus
     val exception = Valid(new LqWriteBundle)
   })
-  
+
   /******************************************************************
    * Structure
    ******************************************************************/
@@ -331,13 +331,13 @@ class LoadQueueUncache(implicit p: Parameters) extends XSModule
 
   /******************************************************************
    * Enqueue
-   * 
+   *
    * s1: hold
    * s2: confirm enqueue and write entry
    *    valid: no redirect, no exception, no replay, is mmio/nc
    *    ready: freelist can allocate
    ******************************************************************/
-  
+
   val s1_sortedVec = HwSort(VecInit(io.req.map { case x => DataWithPtr(x.valid, x.bits, x.bits.uop.robIdx) }))
   val s1_req = VecInit(s1_sortedVec.map(_.bits))
   val s1_valid = VecInit(s1_sortedVec.map(_.valid))
@@ -381,17 +381,17 @@ class LoadQueueUncache(implicit p: Parameters) extends XSModule
 
   /******************************************************************
    * Uncache Transaction
-   * 
+   *
    * 1. uncache req
    * 2. uncache resp
    * 3. writeback
    ******************************************************************/
   private val NC_WB_MOD = NCWBPorts.length
-  
+
   val uncacheReq = Wire(DecoupledIO(io.uncache.req.bits.cloneType))
   val mmioSelect = entries.map(e => e.io.mmioSelect).reduce(_ || _)
   val mmioReq = Wire(DecoupledIO(io.uncache.req.bits.cloneType))
-  // TODO lyq: It's best to choose in robIdx order / the order in which they enter 
+  // TODO lyq: It's best to choose in robIdx order / the order in which they enter
   val ncReqArb = Module(new RRArbiterInit(io.uncache.req.bits.cloneType, LoadUncacheBufferSize))
 
   val mmioOut = Wire(DecoupledIO(io.mmioOut(0).bits.cloneType))
@@ -523,16 +523,16 @@ class LoadQueueUncache(implicit p: Parameters) extends XSModule
 
   /******************************************************************
    * Uncache rollback detection
-   * 
+   *
    * When uncache loads enqueue, it searches uncache loads, They can not enqueue and need re-execution.
-   * 
+   *
    * Cycle 0: uncache enqueue.
    * Cycle 1: Select oldest uncache loads.
    * Cycle 2: Redirect Fire.
    *   Choose the oldest load from LoadPipelineWidth oldest loads.
    *   Prepare redirect request according to the detected rejection.
    *   Fire redirect request (if valid)
-   * 
+   *
    *               Load_S3  .... Load_S3
    * stage 0:        lq            lq
    *                 |             | (can not enqueue)
@@ -543,7 +543,7 @@ class LoadQueueUncache(implicit p: Parameters) extends XSModule
    * stage 2:               lq
    *                        |
    *                     rollback req
-   * 
+   *
    ******************************************************************/
   def selectOldestRedirect(xs: Seq[Valid[Redirect]]): Vec[Bool] = {
     val compareVec = (0 until xs.length).map(i => (0 until i).map(j => isAfter(xs(j).bits.robIdx, xs(i).bits.robIdx)))
